+++ docs/supported-libraries.md
@@ -1,13 +1,13 @@
 ## Supported libraries
 
 | Plugin Name  | Repository Url                             | Min Supported Version | Max Supported Version |
-| ------------ | ------------------------------------------ | --------------------- | --------------------- |
+|--------------| ------------------------------------------ | --------------------- |-----------------------|
 | database/sql | https://pkg.go.dev/database/sql            | -                     | -                     |
 | echo         | https://github.com/labstack/echo           | v4.0.0                | v4.12.0               |
 | fasthttp     | https://github.com/valyala/fasthttp        | v1.45.0               | v1.57.0               |
 | gin          | https://github.com/gin-gonic/gin           | v1.7.0                | v1.10.0               |
 | go-redis     | https://github.com/redis/go-redis          | v9.0.5                | v9.5.1                |
-| go-redis v8  | https://github.com/redis/go-redis          | v8.11.0                | v8.11.5              |
+| go-redis v8  | https://github.com/redis/go-redis          | v8.11.0               | v8.11.5               |
 | gorm         | https://github.com/go-gorm/gorm            | v1.22.0               | v1.25.9               |
 | grpc         | https://google.golang.org/grpc             | v1.44.0               | v1.67.0               |
 | hertz        | https://github.com/cloudwego/hertz         | v0.8.0                | v0.9.2                |
@@ -20,4 +20,5 @@
 | redigo       | https://github.com/gomodule/redigo         | v1.9.0                | v1.9.2                |
 | slog         | https://pkg.go.dev/log/slog                | -                     | -                     |
 | zap          | https://github.com/uber-go/zap             | v1.20.0               | v1.27.0               |
+| fiber        | https://github.com/gofiber/fiber           | v2.43.0               | v2.52.5               |
