+++ pkg/rules/fiberv2/fiberv2_data_type.go
