+++ test/fiber_tests.go
@@ -28,13 +28,13 @@ func init() {
 }
 
 func TestBasicFiberv2(t *testing.T, env ...string) {
-	UseApp("fiberv2/v2.52.5")
+	UseApp("fiberv2/v2.43.0")
 	RunInstrument(t, "-debuglog", "go", "build", "fiber_http.go")
 	RunApp(t, "fiber_http", env...)
 }
 
 func TestBasicFiberv2Https(t *testing.T, env ...string) {
-	UseApp("fiberv2/v2.52.5")
+	UseApp("fiberv2/v2.43.0")
 	RunInstrument(t, "-debuglog", "go", "build", "fiber_https.go")
 	RunApp(t, "fiber_https", env...)
 }
