+++ test/fiberv2/v2.43.0/fiber_http.go
@@ -15,8 +15,6 @@
 package main
 
 import (
-	"crypto/tls"
-	"encoding/json"
 	"fmt"
 	"github.com/alibaba/opentelemetry-go-auto-instrumentation/test/verifier"
 	"go.opentelemetry.io/otel/sdk/trace/tracetest"
@@ -27,14 +25,10 @@ import (
 	"github.com/valyala/fasthttp"
 )
 
-func requestHttpsServer() {
-	client := &fasthttp.Client{
-		TLSConfig: &tls.Config{
-			InsecureSkipVerify: true,
-		},
-	}
+func requestServer() {
+	client := &fasthttp.Client{}
 
-	reqURL := "https://localhost:3000/fiber"
+	reqURL := "http://localhost:3000/fiber"
 
 	req := fasthttp.AcquireRequest()
 	resp := fasthttp.AcquireResponse()
@@ -55,7 +49,7 @@ func requestHttpsServer() {
 	fmt.Printf("Response body is:\n%s", resp.Body())
 }
 
-func setupHttps() {
+func setupHttp() {
 	// Initialize a new Fiber app
 	app := fiber.New()
 
@@ -66,20 +60,18 @@ func setupHttps() {
 	})
 
 	// Start the server on port 3000
-	log.Fatal(app.ListenTLS(":3000", "server.crt", "server.key"))
+	log.Fatal(app.Listen(":3000"))
 }
 
 func main() {
 	// starter server
-	go setupHttps()
+	go setupHttp()
 	time.Sleep(3 * time.Second)
 	// use a http client to request to the server
-	requestHttpsServer()
+	requestServer()
 	// verify trace
 	verifier.WaitAndAssertTraces(func(stubs []tracetest.SpanStubs) {
-		xx, _ := json.Marshal(stubs)
-		fmt.Println(string(xx))
-		verifier.VerifyHttpClientAttributes(stubs[0][0], "GET", "GET", "https://localhost:3000/fiber", "https", "", "tcp", "ipv4", "", "localhost:3000", 200, 0, 3000)
-		verifier.VerifyHttpServerAttributes(stubs[0][1], "GET /fiber", "GET", "https", "tcp", "ipv4", "", "localhost:3000", "fasthttp", "https", "/fiber", "", "/fiber", 200)
+		verifier.VerifyHttpClientAttributes(stubs[0][0], "GET", "GET", "http://localhost:3000/fiber", "http", "", "tcp", "ipv4", "", "localhost:3000", 200, 0, 3000)
+		verifier.VerifyHttpServerAttributes(stubs[0][1], "GET /fiber", "GET", "http", "tcp", "ipv4", "", "localhost:3000", "fasthttp", "http", "/fiber", "", "/fiber", 200)
 	}, 1)
 }
